# src/Api/Controllers/TaskController.cs
using Application.Tasks.DTOs;
using Application.Tasks.Interfaces;
using Domain.Entities.Tasks;
using Domain.Interfaces;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

namespace Api.Controllers;

[Authorize]
[ApiController]
[Route("api/tasks")]
public class TaskController(ITaskService taskService, ITenantContext tenantContext) : ControllerBase
{
    [HttpGet("{id:guid}")]
    public async Task<IActionResult> GetById(Guid id, CancellationToken ct)
    {
        var task = await taskService.GetByIdAsync(tenantContext.TenantId, id, ct);
        if (task is null)
            return NotFound(new { error = $"Task with ID {id} not found." });

        return Ok(task);
    }

    [HttpGet("by-project/{projectId:guid}")]
    public async Task<IActionResult> GetByProjectId(Guid projectId, CancellationToken ct)
    {
        var tasks = await taskService.GetByProjectIdAsync(tenantContext.TenantId, projectId, ct);
        return Ok(tasks);
    }

    [HttpGet("by-assignee/{assigneeId}")]
    public async Task<IActionResult> GetByAssigneeId(Guid assigneeId, CancellationToken ct)
    {
        var tasks = await taskService.GetByAssigneeIdAsync(tenantContext.TenantId, assigneeId, ct);
        return Ok(tasks);
    }

    [HttpGet("by-reporter/{reporterId}")]
    public async Task<IActionResult> GetByReporterId(Guid reporterId, CancellationToken ct)
    {
        var tasks = await taskService.GetByReporterIdAsync(tenantContext.TenantId, reporterId, ct);
        return Ok(tasks);
    }

    [HttpGet("by-status/{status}")]
    public async Task<IActionResult> GetByStatus(TaskItemStatus status, CancellationToken ct)
    {
        var tasks = await taskService.GetByStatusAsync(tenantContext.TenantId, status, ct);
        return Ok(tasks);
    }

    [HttpGet("by-priority/{priority}")]
    public async Task<IActionResult> GetByPriority(TaskPriority priority, CancellationToken ct)
    {
        var tasks = await taskService.GetByPriorityAsync(tenantContext.TenantId, priority, ct);
        return Ok(tasks);
    }

    [HttpPost]
    public async Task<IActionResult> Create([FromBody] CreateTaskDto dto, CancellationToken ct)
    {
        var userId = User.GetUserId();
        await taskService.AddAsync(tenantContext.TenantId, userId, dto, ct);
        return Ok();
    }

    [HttpPatch("{id}")]
    public async Task<IActionResult> Update(Guid id, [FromBody] UpdateTaskDto dto, CancellationToken ct)
    {
        var userId = User.GetUserId();
        await taskService.UpdateAsync(dto, id, tenantContext.TenantId, userId, ct);
        return Ok();
    }

    [HttpDelete("{id}")]
    public async Task<IActionResult> Delete(Guid id, CancellationToken ct)
    {
        var userId = User.GetUserId();
        await taskService.DeleteAsync(tenantContext.TenantId, id, userId, ct);
        return Ok();
    }
}